Preliminary version for the support of async operations. Async requests can be sent and the responses are dispatched to the right caller. For now, no processing.
A small bugfix in XrdClientVector

+++ src/XrdClient/XrdClientConnMgr.hh
@@ -61,23 +61,27 @@ protected:
 public:
    virtual ~XrdClientConnectionMgr();
 
-   short         Connect(XrdClientUrlInfo RemoteAddress);
+   int           Connect(XrdClientUrlInfo RemoteAddress);
    void          Disconnect(short LogConnectionID, bool ForcePhysicalDisc);
-   XrdClientLogConnection *GetConnection(short LogConnectionID);
+
+   XrdClientLogConnection 
+                 *GetConnection(short LogConnectionID);
+
    short         GetPhyConnectionRefCount(XrdClientPhyConnection *PhyConn);
 
-   XrdClientMessage*   ReadMsg(short LogConnectionID);
+   XrdClientMessage*   
+                 ReadMsg(short LogConnectionID);
 
    int           ReadRaw(short LogConnectionID, void *buffer, int BufferLength);
    int           WriteRaw(short LogConnectionID, const void *buffer, 
                           int BufferLength);
 
    static XrdClientConnectionMgr* Instance();
-   static bool IsAlive() {
-     return (fgInstance);
+   static bool   IsAlive() {
+      return (fgInstance);
    }
 
-   static void              Reset();
+   static void   Reset();
 
 
 };
